fixes to download script

+++ scripts/download.py
@@ -5,6 +5,8 @@ import gzip
 import tarfile 
 import shutil
 import urllib.request 
+import urllib 
+import glob
 import warnings 
 warnings.simplefilter('ignore') # Turn off annoying tarfile warnings
 
@@ -13,45 +15,54 @@ warnings.simplefilter('ignore') # Turn off annoying tarfile warnings
 #   ./annotations/pfam
 #   ./annotations/kegg
 #   ./proteins/amino_acids/
-#       ./bacteria
-#       ./archaea
 #   ./proteins/nucleotides/
-#       ./bacteria
-#       ./archaea
 #   ./metadata/
 
 
+def get_latest(dir_path:str) -> str:
+    '''Get the most recently-created file in the specified directory. Returns a complete path, 
+    not just the filename.'''
+    items = glob.glob(os.path.join(dir_path, '*')) # Get all files and subdirectories in the directory. 
+    latest_item = max(items, key=os.path.getctime)
+    return os.path.join(dir_path, latest_item)
 
-def extract_tar(tar_path:str=None, dst_path:str=None, src_path:str=None): 
-    '''
+
+def extract_tar(tar_path:str=None, dst_path:str=None): # , src_path:str=None): 
+    '''Extract a zipped tar file to the specified directory. 
+
     :param tar_path: The path to the tar archive to extract. 
     :param dst_path: The path where all the contents of the tar archive will be written. 
-    :param src_path: The name of the directory where the extracted tar archive is stored. I needed to set
-        this manually. 
     '''
     print(f'extract_tar: Extracting tar archive {tar_path}')
 
     with tarfile.open(tar_path, 'r') as tar:
+        # Extract the tar archive into the same directory. 
         tar.extractall(path=os.path.dirname(tar_path))
-    # src_path is the path to the newly-extracted directory. I can't come up with a good way to automatically
-    # detect what it's name is, so just set it manually. 
-    print(f'extract_tar: Moving files from {src_path} to {dst_path}.')
-    for root, dirs, files in os.walk(src_path):   
-        for file in files:     
-            shutil.move(os.path.join(root, file), os.path.join(dst_path, file))
-            # All the files in the tar archive are gzipped. 
-            if file.split('.')[-1] == 'gz':
-                extract_gz(os.path.join(dst_path, file), dst_path=dst_path, rm=True, verbose=False)
 
-    shutil.rmtree(src_path) # Remove the src_path directory, which should be empty now.
+    # src_path is the path to the newly-extracted file or directory. This should be the most-recently created item in the directory. 
+    src_path = get_latest(os.path.dirname(tar_path, is_directory=True))
+
+    if os.path.isdir(src_path):
+        print(f'extract_tar: Moving files from {src_path} to {dst_path}.')
+        for root, dirs, files in os.walk(src_path):   
+            for file in files:     
+                shutil.move(os.path.join(root, file), os.path.join(dst_path, file))
+                # All the files in the tar archive are gzipped. Need to extract these as well.  
+                if file.split('.')[-1] == 'gz':
+                    extract_gz(os.path.join(dst_path, file), dst_path=dst_path, rm=True, verbose=False)
+        shutil.rmtree(src_path) # Remove the src_path directory, which should be empty now.
+
+    else: # If the extracted item is not a directory, just move it to the destination path. 
+        print(f'extract_tar: Moving file from {src_path} to {dst_path}.')
+        shutil.move(src_path, os.path.join(dst_path, os.path.basename(src_path)))
 
 
-def extract_gz(gz_path:str=None, dst_path:bool=None, rm:bool=True, verbose:bool=True): 
+
+def extract_gz(gz_path:str=None, dst_path:str=None, verbose:bool=True): # , rm:bool=True, 
     '''Extract a zipped file to the specified directory.
     
     :param gz_path: The path to the zipped file. 
     :param dst_path: The path where to which file will be extracted. 
-    :param rm: Whether or not to remove the original file. 
     :param verbose: Whether or not to print a little update. 
     '''
     if verbose: print(f'extract_gz: Extracting gz file {gz_path}')
@@ -63,9 +74,18 @@ def extract_gz(gz_path:str=None, dst_path:bool=None, rm:bool=True, verbose:bool=
     with gzip.open(gz_path, 'rb') as src:
         with open(dst_path, 'wb') as dst:
             shutil.copyfileobj(src, dst)
-    if rm:
-        os.remove(gz_path)
  
+
+def extract(path:str, dst_path:str=None):
+    '''Extract a compressed file, either tar or gz.'''
+    # tar takes precedence over gz. This is because for the non-directories with the .tar.gz file extension, 
+    # the tar utility fully decompresses the file, while gunzip does not.
+    if '.tar' in path:
+        extract_tar(tar_path=path, dst_path=path)
+    elif '.gz' in path:
+        extract_gz(gz_path=path, dst_path=dst_path)
+    else:
+        raise Exception(f'extract: Compressed file at {path} cannot be extracted.')
 
 
 if __name__ == '__main__':
@@ -84,35 +104,42 @@ if __name__ == '__main__':
     url = f'https://data.gtdb.ecogenomic.org/releases/release{args.version}/{args.version}.0/'
 
     # The files we need from GTDB for this are:
-    #   'gtdb_proteins_nt_reps_r{version}.tar.gz'
-    #   'gtdb_proteins_aa_reps_r{version}.tar.gz'
-    #   'ar53_metadata_r{version}.tsv.gz
-    #   'bac120_metadata_r{version}.tsv.gz
     remote_files = [f'genomic_files_reps/gtdb_proteins_nt_reps_r{args.version}.tar.gz']
     remote_files += [f'genomic_files_reps/gtdb_proteins_aa_reps_r{args.version}.tar.gz']
+    # NOTE: In earlier releases, these files are tar archives! Store as tar.gz. So need to try downloading both.
     remote_files += [f'ar53_metadata_r{args.version}.tsv.gz']
     remote_files += [f'bac120_metadata_r{args.version}.tsv.gz']
+    remote_files += [f'ar53_metadata_r{args.version}.tar.gz']
+    remote_files += [f'bac120_metadata_r{args.version}.tar.gz']
 
     local_files = []
     for remote_file in remote_files:
         local_file = remote_file.split('/')[-1]
         local_files.append(local_file)
-        # Download the file if it does not already exist. 
-        if not os.path.exists(os.path.join(data_dir, local_file)):
+        
+        # Skip downloading the file if it does not already exist. 
+        if os.path.exists(os.path.join(data_dir, local_file)):
+            continue
+
+        try:
             print(f'Downloading file from {url + remote_file}')
             urllib.request.urlretrieve(url + remote_file, os.path.join(data_dir, local_file))
-            # local_files.append(wget.download(url + remote_file, out=data_dir))
-            urllib.request.urlretrieve(url + remote_file, os.path.join(data_dir, local_file))
-    
+        except urllib.error.HTTPError:
+            print(f'Failed to download file {remote_file}')
+
+    # Only keep the files in the list which have been succesfully dowloaded. 
+    local_files = [file for file in local_files if os.path.exists(os.path.join(data_dir, file))]
+    assert len(local_files) == 4, f'There should only be 4 files in the local_files list. Found {len(local_files)}.'
+        
     # First, make all necessary directories... 
     os.makedirs(os.path.join(data_dir, 'proteins', 'nucleotides'), exist_ok=True)
     os.makedirs(os.path.join(data_dir, 'proteins', 'amino_acids'), exist_ok=True)
     os.makedirs(os.path.join(data_dir, 'metadata'), exist_ok=True)
     
-    extract_tar(tar_path=os.path.join(data_dir, local_files[0]), dst_path=os.path.join(data_dir, 'proteins', 'nucleotides'), src_path=os.path.join(data_dir, 'protein_fna_reps'))
-    extract_tar(tar_path=os.path.join(data_dir, local_files[1]), dst_path=os.path.join(data_dir, 'proteins', 'amino_acids'), src_path=os.path.join(data_dir, 'protein_faa_reps'))
-    extract_gz(gz_path=os.path.join(data_dir, local_files[2]), dst_path=os.path.join(data_dir, 'metadata'), rm=False)
-    extract_gz(gz_path=os.path.join(data_dir, local_files[3]), dst_path=os.path.join(data_dir, 'metadata'), rm=False)
+    extract(os.path.join(data_dir, local_files[0]), dst_path=os.path.join(data_dir, 'proteins', 'nucleotides'))
+    extract(os.path.join(data_dir, local_files[1]), dst_path=os.path.join(data_dir, 'proteins', 'amino_acids'))
+    extract(os.path.join(data_dir, local_files[2]), dst_path=os.path.join(data_dir, 'metadata'))
+    extract(os.path.join(data_dir, local_files[3]), dst_path=os.path.join(data_dir, 'metadata'))
